+++ eslint.config.mjs
@@ -1,0 +1,3 @@
+import config from '@applet/config/eslint';
+
+export default config;